ai-service/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.9
requests==2.31.0
PyJWT==2.8.0
# torch installed separately in Dockerfile (CPU-only, no gcc needed)
sentence-transformers==2.7.0
huggingface_hub==0.23.4
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0
neo4j==5.14.1
google-generativeai==0.7.2
